use errgroup and handle panic in routine
